refactor shared data packs
